# src/controllers/authController.js
import { registerUserService,loginUserService } from '../services/authService.js';

/**
 * @openapi
 * /api/auth/health:
 *   get:
 *     tags:
 *       - Health
 *     summary: Check if the API is working
 *     description: Simple health check endpoint to verify API is up and running.
 *     responses:
 *       200:
 *         description: API is healthy
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 message:
 *                   type: string
 *                   example: API is working good
 *                 status:
 *                   type: intger
 *                   example: 200
 */
export const CheckHealth = (req, res) => {
  console.log('api is working');
  return res.status(200).json({
    message: 'API is working good',
    status: 200,
  });
};

/**
 * @openapi
 * /api/auth/register:
 *   post:
 *     tags:
 *       - User
 *     summary: Register a new user
 *     description: Create a new user account with username, email, fullname, password, and optional role.
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - username
 *               - email
 *               - fullname
 *               - password
 *             properties:
 *               username:
 *                 type: string
 *                 example: mukesh dhadhariya
 *               email:
 *                 type: string
 *                 example: mukeshdhadhariya1@gmail.com
 *               fullname:
 *                 type: string
 *                 example: Mukesh Dhadhariya
 *               password:
 *                 type: string
 *                 example: password123
 *               role:
 *                 type: string
 *                 description: Optional Role ID reference
 *                 example: 64f3c2a7e1f0b2a7b1c0d123
 *     responses:
 *       201:
 *         description: User registered successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 message:
 *                   type: string
 *                   example: User registered successfully
 *       400:
 *         description: Bad request (missing fields or username/email already exists)
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 message:
 *                   type: string
 *                   example: All fields (username, email, fullname, password) are required
 *       500:
 *         description: Registration failed due to server error
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 message:
 *                   type: string
 *                   example: Registration failed
 *                 error:
 *                   type: string
 *                   example: Error message
 */
export const registerUser = async (req, res) => {
  try {
    const userData = await registerUserService(req.body);
    return res.status(201).json({
      success: true,
      message: 'User registered successfully',
      user: userData
    });
  } catch (error) {
    console.error('Error in registerUser:', error.message);
    res.status(400).json({ success: false, message: error.message });
  }
};

/**
 * @openapi
 * /api/auth/login:
 *   post:
 *     tags:
 *       - Auth
 *     summary: User login
 *     description: Authenticate a user with email and password and return JWT token along with refresh token.
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - email
 *               - password
 *             properties:
 *               email:
 *                 type: string
 *                 format: email
 *                 example: mukesh@example.com
 *               password:
 *                 type: string
 *                 example: yourpassword123
 *     responses:
 *       200:
 *         description: Login successful
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                   example: true
 *                 message:
 *                   type: string
 *                   example: Login successful
 *                 token:
 *                   type: string
 *                   example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
 *                 refreshToken:
 *                   type: string
 *                   example: "d1f2e3c4-5678-90ab-cdef-1234567890ab"
 *       400:
 *         description: Missing email or password
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                   example: false
 *                 message:
 *                   type: string
 *                   example: Email and password are required
 *       401:
 *         description: Authentication failed (invalid credentials)
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                   example: false
 *                 message:
 *                   type: string
 *                   example: Authentication failed
 */
export const loginUser = async (req, res) => {
  try {
    const { email, password } = req.body;
    const result = await loginUserService({ email, password });

    return res.status(200).json({
      success: true,
      message: 'Login successful',
      token: result.accessToken,
      user: result.user,
    });
  } catch (error) {
    console.error('Error in loginUser:', error);
    const status = error.statusCode || 400;
    return res.status(status).json({ success: false, message: error.message || 'Login failed' });
  }
};