# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RBAC API Documentation",
    "version": "1.0.0",
    "description": "Open Source Hackathon (Opcode) project — Role-Based Access Control (RBAC) backend API documentation.\n\nThis documentation provides details of all available API endpoints, authentication methods, and data models used in the RBAC system.",
    "contact": {
      "name": "Mukesh Dhadhariya",
      "email": "mukeshdhadhariya1@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Check if the API is working",
        "description": "Simple health check endpoint to verify API is up and running.",
        "responses": {
          "200": {
            "description": "API is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "API is working good"
                    },
                    "status": {
                      "type": "intger",
                      "example": 200
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Register a new user",
        "description": "Create a new user account with username, email, fullname, password, and optional role.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "email",
                  "fullname",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "mukesh dhadhariya"
                  },
                  "email": {
                    "type": "string",
                    "example": "mukeshdhadhariya1@gmail.com"
                  },
                  "fullname": {
                    "type": "string",
                    "example": "Mukesh Dhadhariya"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  },
                  "role": {
                    "type": "string",
                    "description": "Optional Role ID reference",
                    "example": "64f3c2a7e1f0b2a7b1c0d123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User registered successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing fields or username/email already exists)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "All fields (username, email, fullname, password) are required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Registration failed due to server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Registration failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User login",
        "description": "Authenticate a user with email and password and return JWT token along with refresh token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "mukesh@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "yourpassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Login successful"
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "refreshToken": {
                      "type": "string",
                      "example": "d1f2e3c4-5678-90ab-cdef-1234567890ab"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing email or password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Email and password are required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed (invalid credentials)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Authentication failed"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/permissions": {
      "post": {
        "summary": "Create a new permission",
        "tags": [
          "Permissions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "create_user"
                  },
                  "description": {
                    "type": "string",
                    "example": "Allows creation of users"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permission created successfully"
          },
          "400": {
            "description": "Invalid request data"
          }
        }
      },
      "get": {
        "summary": "Get all permissions",
        "tags": [
          "Permissions"
        ],
        "responses": {
          "200": {
            "description": "List of permissions"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/permissions/{id}": {
      "get": {
        "summary": "Get a permission by ID",
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Permission ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission details"
          },
          "404": {
            "description": "Permission not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update a permission",
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Permission ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "edit_user"
                  },
                  "description": {
                    "type": "string",
                    "example": "Allows editing of users"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permission updated successfully"
          },
          "400": {
            "description": "Invalid request data"
          },
          "404": {
            "description": "Permission not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a permission",
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Permission ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission deleted successfully"
          },
          "404": {
            "description": "Permission not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/roles": {
      "post": {
        "summary": "Create a new role",
        "tags": [
          "Roles"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "admin"
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "64f8a91d1b9d9c123456789a",
                      "64f8a91d1b9d9c123456789b"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role created successfully"
          },
          "400": {
            "description": "Invalid request data"
          }
        }
      },
      "get": {
        "summary": "Get all roles",
        "tags": [
          "Roles"
        ],
        "responses": {
          "200": {
            "description": "List of all roles with their permissions"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/roles/{id}": {
      "get": {
        "summary": "Get a role by ID",
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Role ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role details"
          },
          "404": {
            "description": "Role not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update a role",
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Role ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "manager"
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "64f8a91d1b9d9c123456789a"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated successfully"
          },
          "400": {
            "description": "Invalid request data"
          },
          "404": {
            "description": "Role not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a role",
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Role ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role deleted successfully"
          },
          "404": {
            "description": "Role not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/roles/{id}/permissions": {
      "put": {
        "summary": "Assign permissions to a role",
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Role ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "permissions"
                ],
                "properties": {
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "64f8a91d1b9d9c123456789a",
                      "64f8a91d1b9d9c123456789b"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permissions assigned successfully"
          },
          "400": {
            "description": "Invalid request data"
          },
          "404": {
            "description": "Role not found"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Permissions",
      "description": "API for managing permissions"
    },
    {
      "name": "Roles",
      "description": "API for managing user roles and their permissions"
    }
  ]
}